requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==11.0
openpyxl==3.1.2
python-docx==1.1.0
pandas==2.2.0
json5==0.9.14
Chardet==5.2.0
openai==1.55.3
python-dotenv==1.0.0
# ML dependencies
scikit-learn==1.6.1
joblib==1.5.0
nltk==3.9.1
# Server dependencies
gunicorn==22.0.0